Pull out Daemon Protocol
Make it reusable, having it in node/daemon was not
making it reusable without all the fluff.
