+++ daemon/pom.xml
@@ -28,6 +28,12 @@
       <version>${project.version}</version>
     </dependency>
     <dependency>
+      <groupId>eu.maveniverse.maven.mimir</groupId>
+      <artifactId>daemon-protocol</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+
+    <dependency>
       <groupId>eu.maveniverse.maven.mimir.node</groupId>
       <artifactId>daemon</artifactId>
       <version>${project.version}</version>
